rename: auth context 파일명 파스칼케이스로 수정
