+++ front/src/components/Layouts/NoneLayout.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '../../context/auth-context';
+import { useAuth } from '../../context/AuthContext';
 import styles from './BaseLayout.module.scss';
 
 const NoneLayout = () => {
